requirements.txt
aws-cdk-lib==2.*
constructs>=10,<11
boto3>=1.34.0
requests>=2.31.0
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
syrupy>=4.6.0
moto[dynamodb,sns,sqs]>=5.0.0
aws-lambda-powertools[all]>=2.30.0
types-boto3[dynamodb,sns,sqs]>=1.34.0
ruff>=0.1.0
black>=23.0.0